fix: rename ServiceCard file component
